+++ src/test.rs
@@ -253,11 +253,11 @@ impl crate::server::Handler for TestServerHandler {
                 name: "name1".to_string(),
                 description: Some("description1".to_string()),
                 kind: TiKind::Token,
-                category: EventCategory::Execution,
+                category: Some(EventCategory::Execution),
                 version: "1.0.0".to_string(),
                 patterns: vec![TiRule {
                     rule_id: 9,
-                    category: EventCategory::Unknown,
+                    category: None,
                     name: "rule1".to_string(),
                     description: Some("description1".to_string()),
                     references: Some(vec!["ref1".to_string()]),