+++ src/types.rs
@@ -110,7 +110,6 @@ pub type TrafficFilterRule = (IpNet, Option<Vec<u16>>, Option<Vec<u16>>);
 #[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq, Serialize)]
 #[repr(u8)]
 pub enum EventCategory {
-    Unknown = 0,
     Reconnaissance = 1,
     InitialAccess = 2,
     Execution = 3,
@@ -139,7 +138,7 @@ pub enum TiKind {
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct TiRule {
     pub rule_id: u32,
-    pub category: EventCategory,
+    pub category: Option<EventCategory>,
     pub name: String,
     pub description: Option<String>,
     pub references: Option<Vec<String>>,
@@ -154,7 +153,7 @@ pub struct Tidb {
     pub name: String,
     pub description: Option<String>,
     pub kind: TiKind,
-    pub category: EventCategory,
+    pub category: Option<EventCategory>,
     pub version: String,
     pub patterns: Vec<TiRule>,
 }